Renamed folders to lowercase.
