+++ src/Mailman.Server/ClientApp/src/App.js
@@ -3,13 +3,13 @@ import { Route } from 'react-router'
 import Layout from './components/Layout'
 import Home from './components/Home'
 import Settings from './components/Settings'
-import { TitlePageContainer } from './components/CreateMerge/TitlePageContainer'
-import { TabSelectionContainer } from './components/CreateMerge/TabSelectionContainer'
-import { HeaderSelectionContainer } from './components/CreateMerge/HeaderSelectionContainer'
-import { ReceiverSelectionContainer } from './components/CreateMerge/ReceiverSelectionContainer'
-import { EmailSubjectContainer } from './components/CreateMerge/EmailSubjectContainer'
-import { EmailBodyContainer } from './components/CreateMerge/EmailBodyContainer'
-import { SavePageContainer } from './components/CreateMerge/SavePageContainer'
+import { TitlePageContainer } from './components/create-merge/TitlePageContainer'
+import { TabSelectionContainer } from './components/create-merge/TabSelectionContainer'
+import { HeaderSelectionContainer } from './components/create-merge/HeaderSelectionContainer'
+import { ReceiverSelectionContainer } from './components/create-merge/ReceiverSelectionContainer'
+import { EmailSubjectContainer } from './components/create-merge/EmailSubjectContainer'
+import { EmailBodyContainer } from './components/create-merge/EmailBodyContainer'
+import { SavePageContainer } from './components/create-merge/SavePageContainer'
 import Counter from './components/Counter'
 import FetchData from './components/FetchData'
