+++ src/Mailman.Server/ClientApp/src/components/create-merge/EmailBodyContainer.js
