+++ src/Mailman.Server/ClientApp/src/components/create-merge/ReceiverSelectionContainer.js
