+++ src/Mailman.Server/ClientApp/src/components/merge-template/InfoCard.js
