+++ src/Mailman.Server/ClientApp/src/components/merge-template/InfoCardSidebar.js
