+++ src/Mailman.Server/ClientApp/src/components/merge-template/InfoCardSidebarContainer.js
