+++ src/Mailman.Server/ClientApp/src/components/merge-template/MenuInput.js
